+++ src/includes/functions.inc.php
@@ -163,19 +163,22 @@ function psm_load_conf() {
 function psm_update_conf($key, $value) {
 	global $db;
 
-	$result = $db->save(
-		PSM_DB_PREFIX.'config',
-		array('value' => $value),
-		array('key' => $key)
-	);
-	// save returns the # rows updated, if 0, key doenst exist yet
-	if($result === 0) {
+	// check if key exists
+	$exists = psm_get_conf($key, false);
+	if($exists === false) {
+		// add new config record
 		$db->save(
 			PSM_DB_PREFIX . 'config',
 			array(
 				'key' => $key,
 				'value' => $value,
 			)
+		);
+	} else {
+		$db->save(
+			PSM_DB_PREFIX.'config',
+			array('value' => $value),
+			array('key' => $key)
 		);
 	}
 	$GLOBALS['sm_config'][$key] = $value;
@@ -270,6 +273,8 @@ function psm_curl_get($href, $header = false, $body = true, $timeout = 10, $add_
 	curl_setopt($ch, CURLOPT_NOBODY, (!$body));
 	curl_setopt($ch, CURLOPT_FOLLOWLOCATION, 1);
 	curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
+	curl_setopt($ch, CURLOPT_SSL_VERIFYHOST, 0);
+	curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, 0);
 	curl_setopt($ch, CURLOPT_CONNECTTIMEOUT, $timeout);
 	curl_setopt($ch, CURLOPT_TIMEOUT, $timeout);
 	curl_setopt($ch, CURLOPT_ENCODING, '');
@@ -335,12 +340,9 @@ function psm_date($time) {
  * Check if an update is available for PHP Server Monitor.
  *
  * Will only check for new version if user turned updates on in config.
- * @global object $db
  * @return boolean
  */
 function psm_update_available() {
-	global $db;
-
 	if(!psm_get_conf('show_update')) {
 		// user does not want updates, fair enough.
 		return false;
@@ -370,15 +372,15 @@ function psm_update_available() {
 }
 
 /**
- * Prepare a new Mailer util.
+ * Prepare a new phpmailer instance.
  *
  * If the from name and email are left blank they will be prefilled from the config.
  * @param string $from_name
  * @param string $from_email
- * @return \psm\Util\Mailer
+ * @return \PHPMailer
  */
 function psm_build_mail($from_name = null, $from_email = null) {
-	$phpmailer = new \psm\Util\Mailer();
+	$phpmailer = new \PHPMailer();
 	$phpmailer->Encoding = "base64";
 	$phpmailer->SMTPDebug = false;
 
@@ -410,8 +412,55 @@ function psm_build_mail($from_name = null, $from_email = null) {
 }
 
 /**
+ * Prepare a new SMS util.
+ *
+ * @return \psm\Txtmsg\TxtmsgInterface
+ */
+function psm_build_sms() {
+	$sms = null;
+
+	// open the right class
+	// not making this any more dynamic, because perhaps some gateways need custom settings (like Mollie)
+	switch(strtolower(psm_get_conf('sms_gateway'))) {
+		case 'mosms':
+			$sms = new \psm\Txtmsg\Mosms();
+			break;
+		case 'smsit':
+			$sms = new \psm\Txtmsg\Smsit();
+			break;
+		case 'inetworx':
+			$sms = new \psm\Txtmsg\Inetworx();
+			break;
+		case 'mollie':
+			$sms = new \psm\Txtmsg\Mollie();
+			$sms->setGateway(1);
+			break;
+		case 'spryng':
+			$sms = new \psm\Txtmsg\Spryng();
+			break;
+		case 'clickatell':
+			$sms = new \psm\Txtmsg\Clickatell();
+			break;
+		case 'textmarketer':
+			$sms = new \psm\Txtmsg\Textmarketer();
+			break;
+		case 'smsglobal':
+			$sms = new \psm\Txtmsg\Smsglobal();
+			break;
+	}
+
+	// copy login information from the config file
+	if($sms) {
+		$sms->setLogin(psm_get_conf('sms_gateway_username'), psm_get_conf('sms_gateway_password'));
+		$sms->setOriginator(psm_get_conf('sms_from'));
+	}
+
+	return $sms;
+}
+
+/**
  * Generate a new link to the current monitor
- * @param array $params key value pairs
+ * @param array|string $params key value pairs or pre-formatted string
  * @param boolean $urlencode urlencode all params?
  * @param boolean $htmlentities use entities in url?
  * @return string
@@ -426,13 +475,17 @@ function psm_build_url($params = array(), $urlencode = true, $htmlentities = tru
 
 	if($params != null) {
 		$url .= '?';
-		$delim = ($htmlentities) ? '&amp;' : '&';
+		if(is_array($params)) {
+			$delim = ($htmlentities) ? '&amp;' : '&';
 
-		foreach($params as $k => $v) {
-			if($urlencode) {
-				$v = urlencode($v);
+			foreach($params as $k => $v) {
+				if($urlencode) {
+					$v = urlencode($v);
+				}
+				$url .= $delim . $k . '=' . $v;
 			}
-			$url .= $delim . $k . '=' . $v;
+		} else {
+			$url .= $params;
 		}
 	}
 
@@ -491,7 +544,9 @@ function psm_is_cli() {
  */
 function pre($arr = null) {
 	echo "<pre>";
-	if ($arr === null) debug_print_backtrace();
+	if ($arr === null) {
+		debug_print_backtrace();
+	}
 	print_r($arr);
 	echo "</pre>";
 }