+++ src/psm/Util/Module/ModalInterface.class.php
@@ -19,37 +19,19 @@
  *
  * @package     phpservermon
  * @author      Pepijn Over <pep@neanderthal-technology.com>
+ * @author      Jérôme Cabanis <jerome@lauraly.com>
  * @copyright   Copyright (c) 2008-2014 Pepijn Over <pep@neanderthal-technology.com>
  * @license     http://www.gnu.org/licenses/gpl.txt GNU GPL v3
  * @version     Release: @package_version@
  * @link        http://www.phpservermonitor.org/
- * @since       phpservermon 3.0.0
  **/
 
-namespace psm\Util;
+namespace psm\Util\Module;
 
-/**
- * PHPMailer is not using namespaces so unable to load files in autoloader.
- */
-require_once(PSM_PATH_VENDOR . '/PHPMailer/class.phpmailer.php');
-require_once(PSM_PATH_VENDOR . '/PHPMailer/class.smtp.php');
+interface ModalInterface {
 
-/**
- * PSM Mailer utility
- *
- * The PHPMailer is an open source lib that can be found in vendor/PHPMailer.
- *
- * @see \PHPMailer
- */
-class Mailer extends \PHPMailer {
+	public function __construct(\psm\Service\Template $tpl);
 
-	/**
-	 * Open new PHPMailer
-	 *
-	 * @param boolean $exceptions
-	 */
-	function __construct($exceptions = false) {
-		parent::__construct($exceptions);
-
-	}
+	public function getModalID();
+	public function createHTML();
 }