+++ src/psm/Util/Updater/StatusNotifier.class.php
@@ -223,39 +223,13 @@ class StatusNotifier {
 			return false;
 		}
 
-		// we have to build an userlist for the log table..
-		$userlist = array();
-
-		// open the right class
-		// not making this any more dynamic, because perhaps some gateways need custom settings (like Mollie)
-		switch(strtolower(psm_get_conf('sms_gateway'))) {
-			case 'mosms':
-				$sms = new \psm\Txtmsg\Mosms();
-				break;
-			case 'inetworx':
-				$sms = new \psm\Txtmsg\Inetworx();
-				break;
-			case 'mollie':
-				$sms = new \psm\Txtmsg\Mollie();
-				$sms->setGateway(1);
-				break;
-			case 'spryng':
-				$sms = new \psm\Txtmsg\Spryng();
-				break;
-			case 'clickatell':
-				$sms = new \psm\Txtmsg\Clickatell();
-				break;
-            case 'textmarketer':
-                $sms = new \psm\Txtmsg\Textmarketer();
-                break;
-			case 'smsglobal':
-				$sms = new \psm\Txtmsg\Smsglobal();
-				break;
+		$sms = psm_build_sms();
+		if(!$sms) {
+			return false;
 		}
 
-		// copy login information from the config file
-		$sms->setLogin(psm_get_conf('sms_gateway_username'), psm_get_conf('sms_gateway_password'));
-		$sms->setOriginator(psm_get_conf('sms_from'));
+		// we have to build an userlist for the log table..
+		$userlist = array();
 
 		// add all users to the recipients list
 		foreach ($users as $user) {